+++ BALSAMIC/constants/rules.py
@@ -2,6 +2,12 @@
 from typing import Dict, List
 
 from BALSAMIC.constants.cache import GenomeVersion
+from BALSAMIC.constants.analysis import (
+    AnalysisType,
+    AnalysisWorkflow,
+    SequencingType,
+    WorkflowSolution,
+)
 
 common_cache_rules: List[str] = [
     "snakemake_rules/cache/singularity_containers.rule",
@@ -21,40 +27,45 @@ canfam_cache_rules: List[str] = common_cache_rules + [
     "snakemake_rules/cache/refseq_canfam.rule"
 ]
 
+
 SNAKEMAKE_RULES: Dict[str, Dict[str, list]] = {
     "common": {
-        "concatenate": ["snakemake_rules/concatenation/concatenation.rule"],
         "qc": [
-            "snakemake_rules/quality_control/fastp.rule",
             "snakemake_rules/quality_control/fastqc.rule",
             "snakemake_rules/quality_control/multiqc.rule",
             "snakemake_rules/quality_control/qc_metrics.rule",
-            "snakemake_rules/variant_calling/mergetype_tumor.rule",
+            "snakemake_rules/quality_control/samtools_qc.rule",
         ],
-        "align": [],
+        "align": [
+            "snakemake_rules/align/sentieon_alignment.rule",
+            "snakemake_rules/align/bam_compress.rule",
+        ],
         "varcall": [
             "snakemake_rules/variant_calling/germline_sv.rule",
             "snakemake_rules/variant_calling/sentieon_quality_filter.rule",
             "snakemake_rules/variant_calling/somatic_sv_quality_filter.rule",
         ],
         "annotate": [
-            "snakemake_rules/annotation/vep.rule",
+            "snakemake_rules/annotation/somatic_snv_annotation.rule",
+            "snakemake_rules/annotation/somatic_sv_annotation.rule",
+            "snakemake_rules/annotation/somatic_computations.rule",
+            "snakemake_rules/annotation/germline_annotation.rule",
             "snakemake_rules/annotation/varcaller_sv_filter.rule",
             "snakemake_rules/annotation/vcf2cytosure_convert.rule",
         ],
     },
     "single_targeted": {
         "qc": [
-            "snakemake_rules/quality_control/GATK.rule",
+            "snakemake_rules/quality_control/fastp_tga.rule",
             "snakemake_rules/quality_control/picard.rule",
             "snakemake_rules/quality_control/sambamba_depth.rule",
             "snakemake_rules/quality_control/mosdepth.rule",
+            "snakemake_rules/umi/concatenation_umi.rule",
             "snakemake_rules/umi/qc_umi.rule",
             "snakemake_rules/umi/mergetype_tumor_umi.rule",
             "snakemake_rules/umi/generate_AF_tables.rule",
         ],
         "align": [
-            "snakemake_rules/align/bwa_mem.rule",
             "snakemake_rules/umi/sentieon_umiextract.rule",
             "snakemake_rules/umi/sentieon_consensuscall.rule",
         ],
@@ -73,20 +84,19 @@ SNAKEMAKE_RULES: Dict[str, Dict[str, list]] = {
     },
     "paired_targeted": {
         "qc": [
-            "snakemake_rules/quality_control/GATK.rule",
+            "snakemake_rules/quality_control/fastp_tga.rule",
             "snakemake_rules/quality_control/picard.rule",
             "snakemake_rules/quality_control/sambamba_depth.rule",
             "snakemake_rules/quality_control/mosdepth.rule",
             "snakemake_rules/umi/qc_umi.rule",
-            "snakemake_rules/variant_calling/mergetype_normal.rule",
             "snakemake_rules/quality_control/somalier.rule",
+            "snakemake_rules/umi/concatenation_umi.rule",
             "snakemake_rules/umi/mergetype_tumor_umi.rule",
             "snakemake_rules/umi/mergetype_normal_umi.rule",
             "snakemake_rules/quality_control/contest.rule",
             "snakemake_rules/umi/generate_AF_tables.rule",
         ],
         "align": [
-            "snakemake_rules/align/bwa_mem.rule",
             "snakemake_rules/umi/sentieon_umiextract.rule",
             "snakemake_rules/umi/sentieon_consensuscall.rule",
         ],
@@ -106,11 +116,11 @@ SNAKEMAKE_RULES: Dict[str, Dict[str, list]] = {
     },
     "single_wgs": {
         "qc": [
+            "snakemake_rules/quality_control/fastp_wgs.rule",
             "snakemake_rules/quality_control/sentieon_qc_metrics.rule",
             "snakemake_rules/quality_control/picard_wgs.rule",
             "snakemake_rules/quality_control/report.rule",
         ],
-        "align": ["snakemake_rules/align/sentieon_alignment.rule"],
         "varcall": [
             "snakemake_rules/variant_calling/sentieon_germline.rule",
             "snakemake_rules/variant_calling/sentieon_split_snv_sv.rule",
@@ -124,13 +134,12 @@ SNAKEMAKE_RULES: Dict[str, Dict[str, list]] = {
     },
     "paired_wgs": {
         "qc": [
+            "snakemake_rules/quality_control/fastp_wgs.rule",
             "snakemake_rules/quality_control/sentieon_qc_metrics.rule",
             "snakemake_rules/quality_control/picard_wgs.rule",
             "snakemake_rules/quality_control/report.rule",
-            "snakemake_rules/variant_calling/mergetype_normal.rule",
             "snakemake_rules/quality_control/somalier.rule",
         ],
-        "align": ["snakemake_rules/align/sentieon_alignment.rule"],
         "varcall": [
             "snakemake_rules/variant_calling/sentieon_germline.rule",
             "snakemake_rules/variant_calling/sentieon_split_snv_sv.rule",
@@ -156,10 +165,10 @@ DELIVERY_RULES: List[str] = [
     "collect_custom_qc_metrics",
     "cnv_report",
     # Alignment
-    "mergeBam_tumor",
-    "mergeBam_normal",
     "mergeBam_tumor_umiconsensus",
     "mergeBam_normal_umiconsensus",
+    "bam_compress_tumor",
+    "bam_compress_normal",
     # Germline
     "vep_germline_tumor",
     "vep_germline_normal",