+++ BALSAMIC/models/analysis.py
@@ -1,10 +1,12 @@
+import glob
 import logging
+import os
 import re
 from datetime import datetime
 from pathlib import Path
 from typing import Optional, List, Dict
 
-from pydantic import BaseModel, validator, Field
+from pydantic import BaseModel, validator, Field, root_validator
 from pydantic.types import DirectoryPath, FilePath
 
 from BALSAMIC import __version__ as balsamic_version
@@ -17,6 +19,7 @@ from BALSAMIC.constants.analysis import (
     MutationOrigin,
     MutationType,
     WorkflowSolution,
+    FastqName,
 )
 
 LOG = logging.getLogger(__name__)
@@ -248,9 +251,10 @@ class AnalysisModel(BaseModel):
     log: Optional[DirectoryPath]
     result: Optional[DirectoryPath]
     benchmark: Optional[DirectoryPath]
-    dag: Optional[FilePath]
+    dag: Optional[str]
     BALSAMIC_version: str = balsamic_version
     config_creation_date: Optional[str]
+    pon_version: Optional[str]
 
     class Config:
         validate_all = True
@@ -339,19 +343,11 @@ class AnalysisModel(BaseModel):
     def datetime_as_string(cls, value):
         return datetime.now().strftime("%Y-%m-%d %H:%M")
 
-
-class AnalysisPonModel(AnalysisModel):
-    """Pydantic model containing PON workflow variables
-
-    Attributes:
-        pon_version: Field(str); version of the PON generated file
-    """
-
-    pon_version: str
-
     @validator("pon_version")
     def validate_pon_version(cls, value):
         """Checks that the version matches the following syntax: v<int>"""
+        if value is None:
+            return value
 
         match = re.fullmatch("^v[1-9]\d*$", value)
         if not match:
@@ -362,23 +358,29 @@ class AnalysisPonModel(AnalysisModel):
         return value
 
 
+class FastqInfoModel(BaseModel):
+    """Holds filepaths for forward and reverse reads for a fastq_pattern."""
+
+    fwd: FilePath
+    rev: FilePath
+
+    @validator("fwd", "rev")
+    def fastq_path_as_abspath_str(cls, value):
+        return Path(value).as_posix()
+
+
 class SampleInstanceModel(BaseModel):
     """Holds attributes for samples used in analysis.
 
     Attributes:
         type: Field(str): sample type [tumor, normal]
+        name: Field(str): sample name
+        fastq_info: Field(dict): fastq patterns: paths to forward and reverse fastqs
     """
 
-    type: str
-
-    @validator("type")
-    def sample_type_literal(cls, value):
-        """Validate balsamic supported sample type."""
-        if value not in set(SampleType):
-            raise ValueError(
-                f"The provided sample type ({value}) is not supported in BALSAMIC"
-            )
-        return value
+    type: SampleType
+    name: str
+    fastq_info: Dict[str, FastqInfoModel]
 
 
 class PanelModel(BaseModel):
@@ -398,92 +400,252 @@ class PanelModel(BaseModel):
     chrom: Optional[List[str]]
     pon_cnn: Optional[FilePath]
 
-    @validator("capture_kit")
-    def path_as_abspath_str(cls, value):
-        return Path(value).resolve().as_posix()
-
-    @validator("pon_cnn")
-    def pon_abspath_as_str(cls, value):
+    @validator("pon_cnn", "capture_kit")
+    def file_abspath_as_str(cls, value):
+        """Converts path to string."""
         if value:
             return Path(value).resolve().as_posix()
         return None
 
 
-class PonBalsamicConfigModel(BaseModel):
-    """Summarizes config models in preparation for export
+class ConfigModel(BaseModel):
+    """
+    Class providing common functions and variables for different balsamic workflows.
 
     Attributes:
         QC : Field(QCmodel); variables relevant for fastq preprocessing and QC
+        samples : Field(List[SampleInstanceModel]); List containing samples submitted for analysis
         reference : Field(Dict); dictionary containing paths to reference genome files
         panel : Field(PanelModel(optional)); variables relevant to PANEL BED if capture kit is used
-        singularity : Field(Path); path to singularity container of BALSAMIC
-        rule_directory : Field(Path(RULE_DIRECTORY)); path where snakemake rules can be found
         bioinfo_tools : Field(dict); dictionary of bioinformatics software and which conda/container they are in
         bioinfo_tools_version : Field(dict); dictionary of bioinformatics software and their versions used for the analysis
-    """
-
-    QC: QCModel
-    analysis: AnalysisPonModel
-    samples: Dict[str, SampleInstanceModel]
-    reference: Dict[str, Path]
-    singularity: DirectoryPath
-    bioinfo_tools: dict
-    bioinfo_tools_version: dict
-    panel: Optional[PanelModel]
-
-    @validator("reference")
-    def abspath_as_str(cls, value):
-        for k, v in value.items():
-            value[k] = Path(v).resolve().as_posix()
-        return value
-
-    @validator("singularity")
-    def transform_path_to_dict(cls, value):
-        return {"image": Path(value).resolve().as_posix()}
-
-
-class BalsamicConfigModel(BaseModel):
-    """Summarizes config models in preparation for export
-
-    Attributes:
-        QC : Field(QCmodel); variables relevant for fastq preprocessing and QC
+        singularity : Field(Dict); path to singularity container of BALSAMIC
         vcf : Field(VCFmodel); variables relevant for variant calling pipeline
-        samples : Field(Dict); dictionary containing samples submitted for analysis
-        reference : Field(Dict); dictionary containing paths to reference genome files
-        panel : Field(PanelModel(optional)); variables relevant to PANEL BED if capture kit is used
-        bioinfo_tools : Field(dict); dictionary of bioinformatics software and which conda/container they are in
-        bioinfo_tools_version : Field(dict); dictionary of bioinformatics software and their versions used for the analysis
-        singularity : Field(Path); path to singularity container of BALSAMIC
         background_variants: Field(Path(optional)); path to BACKGROUND VARIANTS for UMI
-        rule_directory : Field(Path(RULE_DIRECTORY)); path where snakemake rules can be found
+        analysis: Field(AnalysisModel); Pydantic model containing workflow variables
+
+    This class also contains functions that help retrieve sample and file information,
+    facilitating BALSAMIC run operations in Snakemake.
+
+    Functions:
+        - get_all_sample_names: Return all sample names in the analysis.
+        - get_fastq_patterns_by_sample: Return all fastq patterns for given samples.
+        - get_all_fastqs_for_sample: Return all fastqs for a sample.
+        - get_fastq_by_fastq_pattern: Return fastq file path for requested fastq pattern and type.
+        - get_sample_name_by_type: Return sample name for requested sample type.
+        - get_sample_type_by_name: Return sample type for requested sample name.
+        - get_bam_name_per_lane: Return list of bam file names for all fastq patterns of a sample.
+        - get_final_bam_name: Return final bam name for downstream analysis.
     """
 
     QC: QCModel
-    vcf: Optional[VCFModel]
-    analysis: AnalysisModel
-    samples: Dict[str, SampleInstanceModel]
+    samples: List[SampleInstanceModel]
     reference: Dict[str, Path]
-    singularity: DirectoryPath
-    background_variants: Optional[FilePath]
-    bioinfo_tools: dict
-    bioinfo_tools_version: dict
+    singularity: Dict[str, DirectoryPath]
+    bioinfo_tools: Dict
+    bioinfo_tools_version: Dict
     panel: Optional[PanelModel]
+    vcf: Optional[VCFModel]
+    background_variants: Optional[FilePath]
+    analysis: AnalysisModel
 
     @validator("reference")
-    def abspath_as_str(cls, value):
-        for k, v in value.items():
-            value[k] = Path(v).resolve().as_posix()
-        return value
+    def abspath_as_str(cls, reference: Dict[str, Path]):
+        for k, v in reference.items():
+            reference[k] = Path(v).resolve().as_posix()
+        return reference
 
     @validator("singularity")
-    def transform_path_to_dict(cls, value):
-        return {"image": Path(value).resolve().as_posix()}
+    def transform_path_to_dict(cls, singularity: Dict[str, DirectoryPath]):
+        for k, v in singularity.items():
+            singularity[k] = Path(v).resolve().as_posix()
+        return singularity
 
     @validator("background_variants")
-    def fl_abspath_as_str(cls, value):
-        if value:
-            return Path(value).resolve().as_posix()
+    def background_variants_abspath_as_str(cls, background_variants: FilePath):
+        """Converts FilePath to string."""
+        if background_variants:
+            return Path(background_variants).resolve().as_posix()
         return None
+
+    @validator("samples")
+    def no_duplicate_fastq_patterns(cls, samples):
+        """Validate that no duplicate fastq patterns have been assigned in dict."""
+        fastq_pattern_counts = {}
+
+        # Count Fastq pattern occurrence
+        for sample in samples:
+            for fastq_pattern in sample.fastq_info.keys():
+                if fastq_pattern not in fastq_pattern_counts:
+                    fastq_pattern_counts[fastq_pattern] = 1
+                else:
+                    fastq_pattern_counts[fastq_pattern] += 1
+
+        # Look for duplicates
+        duplicates = []
+        for fastq_pattern in fastq_pattern_counts:
+            if fastq_pattern_counts[fastq_pattern] > 1:
+                duplicates.append(fastq_pattern)
+
+        if duplicates:
+            raise ValueError(
+                f"Duplicate FastqPattern(s) found: {', '.join(duplicates)} across multiple samples"
+            )
+
+        return samples
+
+    @root_validator(pre=True)
+    def no_unassigned_fastqs_in_fastq_dir(cls, values):
+        """All fastq files in the supplied fastq-dir must have been assigned to the sample-dict."""
+
+        def get_all_fwd_rev_values(samples) -> List[str]:
+            # Return all fastq files in analysis
+            fwd_rev_values = []
+            for sample in samples:
+                for fastq_pattern in sample["fastq_info"]:
+                    fwd_rev_values.append(
+                        sample["fastq_info"][fastq_pattern][FastqName.FWD]
+                    )
+                    fwd_rev_values.append(
+                        sample["fastq_info"][fastq_pattern][FastqName.REV]
+                    )
+            return fwd_rev_values
+
+        fastq_path = values["analysis"]["fastq_path"]
+
+        # Get a set of all fastq files in fastq-directory
+        fastqs_in_fastq_path = set(glob.glob(f"{fastq_path}/*fastq.gz"))
+
+        # Look for fastqs in sample dict
+        fastqs_assigned = set(get_all_fwd_rev_values(values["samples"]))
+
+        unassigned_fastqs = fastqs_in_fastq_path - fastqs_assigned
+        if unassigned_fastqs:
+            raise ValueError(
+                f"Fastqs in fastq-dir not assigned to sample config: {unassigned_fastqs}"
+            )
+
+        return values
+
+    def get_all_sample_names(self) -> List[str]:
+        """Return all sample names in the analysis."""
+        return [sample.name for sample in self.samples]
+
+    def get_fastq_patterns_by_sample(self, sample_names: List[str]) -> List[str]:
+        """Return all fastq_patterns for a given sample."""
+        return [
+            fastq_pattern
+            for sample in self.samples
+            if sample.name in sample_names
+            for fastq_pattern in sample.fastq_info.keys()
+        ]
+
+    def get_all_fastqs_for_sample(
+        self, sample_name: str, fastq_types: Optional[List[FastqName]] = None
+    ) -> List[str]:
+        """Return all fastqs (optionally only [fwd/rev]) involved in analysis of sample."""
+
+        if fastq_types is None:
+            fastq_types = [FastqName.FWD, FastqName.REV]
+
+        fastq_list: List = []
+        for sample in self.samples:
+            if sample.name == sample_name:
+                for fastq_info in sample.fastq_info.values():
+                    if FastqName.FWD in fastq_types:
+                        fastq_list.append(fastq_info.fwd)
+                    if FastqName.REV in fastq_types:
+                        fastq_list.append(fastq_info.rev)
+        return fastq_list
+
+    def get_all_fastq_names(self, remove_suffix: bool = False) -> List[str]:
+        """Return all fastq_names involved in analysis, optionally remove fastq.gz suffix."""
+        fastq_names = []
+        for sample in self.samples:
+            for fastq_pattern, fastqs in sample.fastq_info.items():
+                if remove_suffix:
+                    fastq_names.extend(
+                        [
+                            Path(fastqs.fwd).name.replace(".fastq.gz", ""),
+                            Path(fastqs.rev).name.replace(".fastq.gz", ""),
+                        ]
+                    )
+                else:
+                    fastq_names.extend(
+                        [
+                            Path(fastqs.fwd).name,
+                            Path(fastqs.rev).name,
+                        ]
+                    )
+        return fastq_names
+
+    def get_fastq_by_fastq_pattern(
+        self, fastq_pattern: str, fastq_type: FastqName
+    ) -> str:
+        """Return fastq file path for requested fastq pair pattern and fastq type: [fwd/rev]."""
+        for sample in self.samples:
+            if fastq_pattern in sample.fastq_info:
+                return (
+                    sample.fastq_info[fastq_pattern].fwd
+                    if fastq_type == FastqName.FWD
+                    else sample.fastq_info[fastq_pattern].rev
+                )
+
+    def get_sample_name_by_type(self, sample_type: str) -> str:
+        """Return sample name for requested sample type."""
+        for sample in self.samples:
+            if sample.type == sample_type:
+                return sample.name
+
+    def get_sample_type_by_name(self, sample_name: str, uppercase: bool = False) -> str:
+        """Return sample type for requested sample name, optionally return it capitalized"""
+        for sample in self.samples:
+            if sample.name == sample_name:
+                return sample.type.upper() if uppercase else sample.type
+
+    def get_bam_name_per_lane(self, bam_dir: str, sample_name: str) -> List[str]:
+        """Return list of bam-file names for all fastq_patterns of a sample."""
+        bam_names = []
+        for sample in self.samples:
+            if sample.name == sample_name:
+                bam_names.extend(
+                    [
+                        f"{bam_dir}{sample_name}_align_sort_{fastq_pattern}.bam"
+                        for fastq_pattern in sample.fastq_info
+                    ]
+                )
+        return bam_names
+
+    def get_final_bam_name(
+        self, bam_dir: str, sample_name: str = None, sample_type: str = None
+    ) -> str:
+        """Return final bam name to be used in downstream analysis."""
+
+        if not sample_name and not sample_type:
+            raise ValueError(
+                "Either sample_name or sample_type must be provided to get the final bam name."
+            )
+
+        sample_name = (
+            self.get_sample_name_by_type(sample_type)
+            if not sample_name
+            else sample_name
+        )
+
+        sample_type = (
+            self.get_sample_type_by_name(sample_name)
+            if not sample_type
+            else sample_type
+        )
+
+        if self.analysis.analysis_type == AnalysisType.PON:
+            # Only dedup is necessary for panel of normals
+            final_bam_suffix = "dedup"
+        else:
+            # For every analysis except PON, the name of the final processed bamfile is defined here
+            final_bam_suffix = "dedup.realign"
+
+        return f"{bam_dir}{sample_type}.{sample_name}.{final_bam_suffix}.bam"
 
 
 class UMIParamsCommon(BaseModel):